[Task-2-1] Reorganized project structure
