+++ vadim.nesmeianov/task-2-1/cmd/comp_op/main.go
