+++ vadim.nesmeianov/task-2-1/cmd/service/main.go
@@ -1,6 +1,6 @@
 package main
 
-import "task-2-1/internal/office"
+import "task-2-1/cmd/office"
 
 func main() {
 	office.Run(true)
